feat: Make HMSS and Trustee produce identical results (#3637)
Closes #3636

+++ src/main/kotlin/org/wfanet/measurement/edpaggregator/resultsfulfiller/KAnonymizer.kt
@@ -54,7 +54,7 @@ object KAnonymizer {
     val reachValue =
       ReachAndFrequencyComputations.computeReach(
         rawHistogram = histogram,
-        vidSamplingIntervalWidth = measurementSpec.vidSamplingInterval.width,
+        vidSamplingIntervalWidth = measurementSpec.vidSamplingInterval.width.toDouble(),
         vectorSize = maxPopulation,
         dpParams = null,
         kAnonymityParams = kAnonymityParams,
